Disabled ulimits propagation to nodes for IB
By default Slurm propagates to the job environment the same ulimits in place on the node where the job has been submitted. 
If this is a VM with a non-HPC image, max locked memory is set to 64 KB. This would prevent IB to pin enough memory for RDMA in the compute node.